requirements.txt
retrying==1.3.3
pandas==0.25.1
scikit-learn==0.21.3
matplotlib==3.1.1
numpy>=1.16.3
tensorflow==2.0.0
joblib==0.13.2
papermill==1.2.0
ray==0.7.5
python-pachyderm==2.0.0
hyperopt==0.1.2
pylib==1.0.0
requests>=2.22.0
schedule==0.6.0
Pillow==6.0.0
azure-storage-blob==2.1.0
# https://github.com/SeldonIO/seldon-core/issues/1023
seldon-core==0.4.1
neural_structured_learning==1.1.0